Bug 1216460 - [1.1] Refactor data types, fix logs and prevent harmful type promotions in SourceBuffer eviction handling. r=jya
UltraBlame original commit: 6495286a5790d50950cb6c2b9afdae0ef1fe4538

+++ dom/media/mediasource/SourceBufferContentManager.h
@@ -70,7 +70,7 @@ public:
   
   virtual EvictDataResult
   EvictData(media::TimeUnit aPlaybackTime,
-            uint32_t aThreshold,
+            int64_t aThreshold,
             media::TimeUnit* aBufferStartTime) = 0;
